Put R_GetProductId() and isTuyaManufacturerName() in own module
